pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Zin Yosrim <zinyosrim@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pandas = "^1.3.5"
azure-storage-blob = "^12.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
